fix(plugin-react-navigation): add full support for RNScreens 4.0-4.15
