+++ .changeset/violet-experts-accept.md
@@ -1,0 +1,5 @@
+---
+"@brandingbrand/code-plugin-react-navigation": minor
+---
+
+Add compatibility for RNScreens >=4.0 <4.16
