+++ packages/plugin-react-navigation/package.json
@@ -24,8 +24,9 @@
     "typescript": "^5.3.3"
   },
   "peerDependencies": {
-    "@brandingbrand/code-cli-kit": "^14.0.0",
+    "@brandingbrand/code-cli-kit": "^14.2.1",
     "@react-navigation/native": "^7",
+    "react-native-screens": "^4",
     "react-native": ">=0.73.0"
   },
   "publishConfig": {
